Form for problem C.
I'm thinking this form could work well for part c.  A user could select the radio button for the maker.  Onchange, a query could be run using the maker value as the input, and a list of items could populate a list of models, either as a part of a combo-box or some other array.  That would be our select statement.  The user could then select the model.  Pressing submit would activate an insert statement into the orders table using the values from form.  Honestly, I think this is a big enough project with out attempting the other problems in the set.  What do you think?